SSCSCI-969: remove reference to SscsOcrDataUtil
