+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/validation/helper/SscsDataHelper.java
@@ -18,6 +18,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.text.CaseUtils;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
@@ -234,4 +235,16 @@ public class SscsDataHelper {
         return null;
     }
 
+    public static boolean findBooleanExists(String... values) {
+        for (String v : values) {
+            if (StringUtils.isNotBlank(v)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    public static String getField(Map<String, Object> pairs, String field) {
+        return pairs != null && pairs.containsKey(field) && pairs.get(field) != null ? StringUtils.trimToNull(pairs.get(field).toString()) : null;
+    }
 }
